# OpenAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "F1 API",
    "description": "An easier way to use the [ergast.com](https://ergast.com/mrd/) F1 API, with correct types!",
    "license": {
      "name": "MIT",
      "url": "https://github.com/niek-o/F1API/blob/main/LICENSE.md"
    },
    "version": "1.5.0",
    "x-logo": {
      "url": "https://upload.wikimedia.org/wikipedia/commons/f/f2/New_era_F1_logo.png"
    }
  },
  "paths": {
...
    "/constructors": {
      "get": {
        "tags": [
          "Constructors"
        ],
        "summary": "Get Constructors",
        "operationId": "get_constructors",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Constructors"
                },
                "example": {
                  "constructors": [
                    {
                      "constructorId": "red_bull",
                      "url": "https://://en.wikipedia.org/wiki/Red_Bull_Racing",
                      "name": "Red Bull",
                      "nationality": "Austrian"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Constructors not found"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/constructors/{season}": {
      "get": {
        "tags": [
          "Constructors",
          "Season"
        ],
        "summary": "Get Constructors By Season",
        "operationId": "get_constructors_by_season",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "string"
            },
            "name": "season",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Constructors"
                },
                "example": {
                  "constructors": [
                    {
                      "constructorId": "red_bull",
                      "url": "https://://en.wikipedia.org/wiki/Red_Bull_Racing",
                      "name": "Red Bull",
                      "nationality": "Austrian"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Constructors not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/constructor/{constructor_id}": {
      "get": {
        "tags": [
          "Constructors"
        ],
        "summary": "Get Constructor By Id",
        "operationId": "get_constructor_by_id",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Constructor Id",
              "type": "string"
            },
            "name": "constructor_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Constructor"
                },
                "example": {
                  "constructorId": "red_bull",
                  "url": "https://://en.wikipedia.org/wiki/Red_Bull_Racing",
                  "name": "Red Bull",
                  "nationality": "Austrian"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Constructor not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/drivers": {
      "get": {
        "tags": [
          "Drivers"
        ],
        "summary": "Get Drivers",
        "operationId": "get_drivers",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drivers"
                },
                "example": {
                  "drivers": [
                    {
                      "driverId": "max_verstappen",
                      "url": "https://://en.wikipedia.org/wiki/Max_Verstappen",
                      "givenName": "Max",
                      "familyName": "Verstappen",
                      "dateOfBirth": "1997-09-30",
                      "nationality": "Dutch",
                      "permanentNumber": 33,
                      "code": "VER"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Drivers not found"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{season}": {
      "get": {
        "tags": [
          "Drivers",
          "Season"
        ],
        "summary": "Get Drivers By Season",
        "operationId": "get_drivers_by_season",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "integer"
            },
            "name": "season",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drivers"
                },
                "example": {
                  "drivers": [
                    {
                      "driverId": "max_verstappen",
                      "url": "https://://en.wikipedia.org/wiki/Max_Verstappen",
                      "givenName": "Max",
                      "familyName": "Verstappen",
                      "dateOfBirth": "1997-09-30",
                      "nationality": "Dutch",
                      "permanentNumber": 33,
                      "code": "VER"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Drivers not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/driver/{driver_id}": {
      "get": {
        "tags": [
          "Drivers"
        ],
        "summary": "Get Driver By Id",
        "operationId": "get_driver_by_id",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "string"
            },
            "name": "driver_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Driver"
                },
                "example": {
                  "driverId": "max_verstappen",
                  "url": "https://://en.wikipedia.org/wiki/Max_Verstappen",
                  "givenName": "Max",
                  "familyName": "Verstappen",
                  "dateOfBirth": "1997-09-30",
                  "nationality": "Dutch",
                  "permanentNumber": 33,
                  "code": "VER"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Driver not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Constructor": {
        "title": "Constructor",
        "required": [
          "constructorId",
          "url",
          "name",
          "nationality"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "constructorId": {
            "title": "Constructorid",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nationality": {
            "title": "Nationality",
            "type": "string"
          }
        }
      },
...
      "Constructors": {
        "title": "Constructors",
        "required": [
          "constructors"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "constructors": {
            "title": "Constructors",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Constructor"
            }
          }
        }
      },
      "Driver": {
        "title": "Driver",
        "required": [
          "driverId",
          "url",
          "givenName",
          "familyName",
          "dateOfBirth",
          "nationality"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "driverId": {
            "title": "Driverid",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "givenName": {
            "title": "Givenname",
            "type": "string"
          },
          "familyName": {
            "title": "Familyname",
            "type": "string"
          },
          "dateOfBirth": {
            "title": "Dateofbirth",
            "type": "string"
          },
          "nationality": {
            "title": "Nationality",
            "type": "string"
          },
          "permanentNumber": {
            "title": "Permanentnumber",
            "type": "integer"
          },
          "code": {
            "title": "Code",
            "type": "string"
          }
        }
      },
...
      "Drivers": {
        "title": "Drivers",
        "required": [
          "drivers"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "drivers": {
            "title": "Drivers",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Driver"
            }
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "Message": {
        "title": "Message",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}